Add rule for one argument hash not using {}
You can't do `m {one: 1, two: 2}` as a method call because ruby expects the braces to be an inline block there and throws a syntax error on ":". Instead you *have* to include the parens, `m({one: 1, two: 2})`, so we might as well drop the `{}` at that point and just go with `m(one: 1, two: 2)` instead. (Or `m one: 1, two: 2` instead if that doesn't make it ambiguous.)