Add explicit note about bare rescues
`rescue` only rescues `StandardError`, but our example in the style guide made it look like it rescued `Exception`. It's also surprising to come across in code if you don't remember that it only rescues `StandardError` and not `Exception`. For that reason we should be explicit and avoid it in our code, preferring `rescue StandardError` or `rescue Exception` to show what we want to rescue.